+++ tests/python/test_household.py
@@ -1,48 +1,17 @@
 import pytest
-from flask import Flask
 import json
 from unittest.mock import MagicMock, patch
 from sqlalchemy.engine.row import LegacyRow
 
 from policyengine_api.routes.household_routes import household_bp
 from policyengine_api.services.household_service import HouseholdService
-from policyengine_api.constants import COUNTRY_PACKAGE_VERSIONS
 
-# TODO: Check if this format is correct
-SAMPLE_HOUSEHOLD_DATA = {
-    "data": {"people": {"person1": {"age": 30, "income": 50000}}},
-    "label": "Test Household",
-}
-
-SAMPLE_DB_ROW = {
-    "id": 1,
-    "country_id": "us",
-    "household_json": json.dumps(SAMPLE_HOUSEHOLD_DATA["data"]),
-    "household_hash": "some-hash",
-    "label": "Test Household",
-    "api_version": "3.0.0",
-}
-
-
-# These will be moved to the correct location once
-# testing PR that creates folder structure is merged
-@pytest.fixture
-def mock_database():
-    """Mock the database module."""
-    with patch(
-        "policyengine_api.services.household_service.database"
-    ) as mock_db:
-        yield mock_db
-
-
-@pytest.fixture
-def mock_hash_object():
-    """Mock the hash_object function."""
-    with patch(
-        "policyengine_api.services.household_service.hash_object"
-    ) as mock:
-        mock.return_value = "some-hash"
-        yield mock
+from tests.fixtures.household_fixtures import (
+    SAMPLE_HOUSEHOLD_DATA,
+    SAMPLE_DB_ROW,
+    mock_database,
+    mock_hash_object,
+)
 
 
 class TestGetHousehold:
